requirements.txt
asgiref==3.5.2
chardet==5.0.0
DateTimeRange==1.2.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
mbstrdecoder==1.1.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typepy==1.3.0
tzdata==2022.1